twilio_security_scanner/requirements.txt
twilio==9.4
python-dotenv==1.0.1